tests/ghostpy/requirements.txt
Ghost.py==0.1b
